+++ PMProject/PMProject/ViewController/FeedViewController/FeedViewController.swift
@@ -24,7 +24,7 @@ class FeedViewController: BalanceProvidingViewController {
     }
 
     // MARK: - BalanceProviderDelegate
-    override func update(balance: Int) {
+    override func update(balance: Double) {
         DispatchQueue.main.async { [weak self] in
             guard let navigationItem = self?.navigationItem else {
                 return
